refactor: group contact-form files by contact-form folder in the internal package
