+++ internal/internal/src/contact-form/types.ts
@@ -1,0 +1,7 @@
+export interface ContactFormDataProps {
+  name: string;
+  email: string;
+  telegram: string;
+  message: string;
+  source: string;
+}
